A touch of slight refactoring
